+++ packages/checkout-extension/src/ExtensionRegionContainer.ts
@@ -1,0 +1,4 @@
+export enum ExtensionRegionContainer {
+    ShippingShippingAddressFormBefore = 'extension-region-shipping-shippingaddressform-before',
+    ShippingShippingAddressFormAfter = 'extension-region-shipping-shippingaddressform-after',
+}