+++ packages/checkout-extension/src/index.ts
@@ -1,0 +1,13 @@
+export { isCheckoutExtensionEnabled } from './isCheckoutExtensionEnabled';
+export { ExtensionRegionContainer } from './ExtensionRegionContainer';
+export { ExtensionService } from './ExtensionService';
+export {
+    ExtensionState,
+    ExtensionAction,
+    ExtensionActionType,
+    ExtensionProvider,
+} from './ExtensionProvider';
+export { ExtensionContextProps, ExtensionContext } from './ExtensionContext';
+export { withExtension } from './withExtension';
+export { getExtensions } from './Extension.mock';
+export { useExtensions } from './useExtensions';