+++ packages/checkout-extension/src/isCheckoutExtensionEnabled.ts
@@ -1,0 +1,7 @@
+import { CheckoutSettings } from '@bigcommerce/checkout-sdk';
+
+export function isCheckoutExtensionEnabled(
+    checkoutSettings: CheckoutSettings | undefined,
+): boolean {
+    return Boolean(checkoutSettings?.features['PROJECT-5029.checkout_extension']);
+}