+++ packages/core/src/app/customer/SubscribeField.tsx
@@ -1,7 +1,7 @@
 import { FieldProps } from 'formik';
 import React, { FunctionComponent, memo } from 'react';
 
-import { TranslatedString } from '@bigcommerce/checkout/locale';
+import { TranslatedHtml } from '@bigcommerce/checkout/locale';
 
 import { Input, Label } from '../ui/form';
 
@@ -19,12 +19,12 @@ const SubscribeField: FunctionComponent<SubscribeFieldProps> = ({
             checked={field.value}
             className="form-checkbox"
             id={field.name}
-            type="checkbox"
             testId="should-subscribe-checkbox"
+            type="checkbox"
         />
 
         <Label htmlFor={field.name}>
-            <TranslatedString
+            <TranslatedHtml
                 id={
                     requiresMarketingConsent
                         ? 'customer.guest_marketing_consent'