+++ packages/core/src/app/shipping/ShippingFormFooter.tsx
@@ -1,5 +1,7 @@
-import React, { PureComponent, ReactNode } from 'react';
+import { ExtensionRegion } from '@bigcommerce/checkout-sdk';
+import React, { FunctionComponent, useEffect } from 'react';
 
+import { ExtensionRegionContainer, useExtensions } from '@bigcommerce/checkout/checkout-extension';
 import { TranslatedString } from '@bigcommerce/checkout/locale';
 
 import { OrderComments } from '../orderComments';
@@ -18,60 +20,78 @@ export interface ShippingFormFooterProps {
     isLoading: boolean;
 }
 
-class ShippingFormFooter extends PureComponent<ShippingFormFooterProps> {
-    render(): ReactNode {
-        const {
-            cartHasChanged,
-            isMultiShippingMode,
-            shouldShowOrderComments,
-            shouldShowShippingOptions = true,
-            shouldDisableSubmit,
-            isLoading,
-        } = this.props;
+const ShippingFormFooter: FunctionComponent<ShippingFormFooterProps> = ({
+    cartHasChanged,
+    isMultiShippingMode,
+    shouldShowOrderComments,
+    shouldShowShippingOptions = true,
+    shouldDisableSubmit,
+    isLoading,
+}) => {
+    const { extensionService, isExtensionEnabled } = useExtensions();
+    const isExtensionRegionEnabled = Boolean(
+        isExtensionEnabled() &&
+            extensionService.isRegionEnabled(ExtensionRegion.ShippingShippingAddressFormAfter),
+    );
 
-        return (
-            <>
-                <Fieldset
-                    id="checkout-shipping-options"
-                    legend={
-                        <>
-                            <Legend>
-                                <TranslatedString id="shipping.shipping_method_label" />
-                            </Legend>
+    useEffect(() => {
+        if (isExtensionRegionEnabled) {
+            void extensionService.renderExtension(
+                ExtensionRegionContainer.ShippingShippingAddressFormAfter,
+                ExtensionRegion.ShippingShippingAddressFormAfter,
+            );
 
-                            {cartHasChanged && (
-                                <Alert type={AlertType.Error}>
-                                    <strong>
-                                        <TranslatedString id="shipping.cart_change_error" />
-                                    </strong>
-                                </Alert>
-                            )}
-                        </>
-                    }
+            return () => {
+                extensionService.removeListeners(ExtensionRegion.ShippingShippingAddressFormAfter);
+            };
+        }
+    }, [extensionService, isExtensionRegionEnabled]);
+
+    return (
+        <>
+            {isExtensionRegionEnabled && (
+                <div id={ExtensionRegionContainer.ShippingShippingAddressFormAfter} />
+            )}
+            <Fieldset
+                id="checkout-shipping-options"
+                legend={
+                    <>
+                        <Legend>
+                            <TranslatedString id="shipping.shipping_method_label" />
+                        </Legend>
+
+                        {cartHasChanged && (
+                            <Alert type={AlertType.Error}>
+                                <strong>
+                                    <TranslatedString id="shipping.cart_change_error" />
+                                </strong>
+                            </Alert>
+                        )}
+                    </>
+                }
+            >
+                <ShippingOptions
+                    isMultiShippingMode={isMultiShippingMode}
+                    isUpdatingAddress={isLoading}
+                    shouldShowShippingOptions={shouldShowShippingOptions}
+                />
+            </Fieldset>
+
+            {shouldShowOrderComments && <OrderComments />}
+
+            <div className="form-actions">
+                <Button
+                    disabled={shouldDisableSubmit}
+                    id="checkout-shipping-continue"
+                    isLoading={isLoading}
+                    type="submit"
+                    variant={ButtonVariant.Primary}
                 >
-                    <ShippingOptions
-                        isMultiShippingMode={isMultiShippingMode}
-                        isUpdatingAddress={isLoading}
-                        shouldShowShippingOptions={shouldShowShippingOptions}
-                    />
-                </Fieldset>
-
-                {shouldShowOrderComments && <OrderComments />}
-
-                <div className="form-actions">
-                    <Button
-                        disabled={shouldDisableSubmit}
-                        id="checkout-shipping-continue"
-                        isLoading={isLoading}
-                        type="submit"
-                        variant={ButtonVariant.Primary}
-                    >
-                        <TranslatedString id="common.continue_action" />
-                    </Button>
-                </div>
-            </>
-        );
-    }
-}
+                    <TranslatedString id="common.continue_action" />
+                </Button>
+            </div>
+        </>
+    );
+};
 
 export default ShippingFormFooter;