+++ packages/paypal-commerce-integration/src/PayPalCommercePaymentMethod.tsx
@@ -9,6 +9,12 @@ import {
 import PayPalCommercePaymentMethodComponent from './components/PayPalCommercePaymentMethodComponent';
 
 const PayPalCommercePaymentMethod: FunctionComponent<PaymentMethodProps> = props => {
+    const isPaymentDataRequired = props.checkoutState.data.isPaymentDataRequired();
+
+    if (!isPaymentDataRequired) {
+        return null;
+    }
+
     return <PayPalCommercePaymentMethodComponent
         providerOptionsKey="paypalcommerce"
         {...props}